# tools/null-g-proxy/openapi.yaml
openapi: 3.0.3
info:
  title: Antigravity Proxy API
  description: |
    OpenAI-compatible HTTP proxy for the Antigravity IDE.
    Exposes chat completions, Git helpers, a knowledge base, terminal execution,
    and code intelligence — all from a single local REST API.
  version: 1.0.0
  contact:
    name: null-g-proxy
  license:
    name: MIT
servers:
  - url: http://localhost:8787
    description: Local development server
...
paths:
...
  /v1/git/commit-message:
    post:
      tags: [Git]
      summary: Generate AI commit message
      description: |
        Uses the Antigravity AI to generate a commit message for the given diff.
        If `diff` is omitted, runs `git diff --staged` in `workingDir`.
      operationId: generateCommitMessage
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitMessageRequest'
            examples:
              auto:
                summary: Auto-detect staged diff
                value:
                  workingDir: /Users/user/workspace/myproject
                  style: conventional
              manual:
                summary: Provide diff manually
                value:
                  style: imperative
                  diff: |
                    diff --git a/src/auth.ts b/src/auth.ts
                    +  const token = jwt.sign(payload, secret, { expiresIn: '15m' });
      responses:
        '200':
          description: Generated commit message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommitMessageResponse'
              example:
                message: "feat(auth): add JWT token refresh endpoint"
                subject: "feat(auth): add JWT token refresh endpoint"
                body: "Implements /api/auth/refresh that issues new access tokens\nusing a valid refresh token stored in httpOnly cookie."
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '502':
          $ref: '#/components/responses/UpstreamError'
...
  /v1/git/worktree:
    post:
      tags: [Git]
      summary: Create a Git worktree
      description: Runs `git worktree add` to create a new worktree for the given branch.
      operationId: createWorktree
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorktreeRequest'
            example:
              repo: myproject
              branch: feature/new-ui
      responses:
        '200':
          description: Worktree created
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string
                    example: /Users/user/workspace/myproject-feature-new-ui
                  branch:
                    type: string
                    example: feature/new-ui
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─── Knowledge ─────────────────────────────────────────────────────────────
...
  /v1/knowledge/search:
    get:
      tags: [Knowledge]
      summary: Search knowledge base
      description: Full-text search across all knowledge items. Name matches are weighted 5×.
      operationId: searchKnowledge
      parameters:
        - name: q
          in: query
          required: true
          description: Search query string
          schema:
            type: string
          example: authentication
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeSearchResponse'
              example:
                query: authentication
                results:
                  - name: auth-notes
                    filename: auth-notes.md
                    path: /Users/user/.gemini/antigravity/knowledge/auth-notes.md
                    size: 512
                    createdAt: 1700000000
                    modifiedAt: 1700000100
                    preview: "# Auth Notes  We use JWT..."
                    content: "# Auth Notes\n\nWe use JWT with 15min expiry..."
                    relevance: 6
                total: 1
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/knowledge/items:
    post:
      tags: [Knowledge]
      summary: Create a knowledge item
      description: |
        Creates a new Markdown file in the global knowledge directory
        (`~/.gemini/antigravity/knowledge/`) or in `<projectDir>/.antigravity/knowledge/`.
        Returns 409 if an item with the same name already exists.
      operationId: createKnowledgeItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateKnowledgeRequest'
            example:
              name: auth-notes
              content: "# Auth Notes\n\nWe use JWT with 15min expiry and refresh tokens."
      responses:
        '201':
          description: Knowledge item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeItem'
              example:
                name: auth-notes
                filename: auth-notes.md
                path: /Users/user/.gemini/antigravity/knowledge/auth-notes.md
                size: 512
                createdAt: 1700000000
                modifiedAt: 1700000000
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Knowledge item already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  message: Knowledge item "auth-notes.md" already exists. Use a different name.
                  type: conflict
                  code: already_exists

  # ─── Terminal ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CommitMessageRequest:
      type: object
      properties:
        diff:
          type: string
          description: Raw unified diff. If omitted, `git diff --staged` is run in workingDir.
        workingDir:
          type: string
          description: Working directory for auto-detecting staged diff (defaults to cwd)
        style:
          type: string
          enum: [conventional, imperative, descriptive]
          default: conventional
          description: Commit message style
    CommitMessageResponse:
      type: object
      properties:
        message:
          type: string
          description: Full commit message (subject + blank line + body)
        subject:
          type: string
          description: First line / subject (max 72 chars)
        body:
          type: string
          description: Optional body with more detail
...
    WorktreeRequest:
      type: object
      required: [repo, branch]
      properties:
        repo:
          type: string
          description: Repo name (relative to ~/workspace) or absolute path
          example: myproject
        branch:
          type: string
          description: Branch name to checkout in the new worktree
          example: feature/new-ui
        path:
          type: string
          description: Explicit path for the worktree (auto-generated if omitted)

    # Knowledge
    KnowledgeItem:
      type: object
      properties:
        name:
          type: string
        filename:
          type: string
        path:
          type: string
        size:
          type: integer
          description: File size in bytes
        createdAt:
          type: integer
          description: Unix timestamp
        modifiedAt:
          type: integer
          description: Unix timestamp
        preview:
          type: string
          description: First 200 chars of content
...
    KnowledgeSearchResult:
      allOf:
        - $ref: '#/components/schemas/KnowledgeItem'
        - type: object
          properties:
            content:
              type: string
              description: Full file content
            relevance:
              type: integer
              description: Match score (name matches weighted 5×)
    KnowledgeSearchResponse:
      type: object
      properties:
        query:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgeSearchResult'
        total:
          type: integer
    CreateKnowledgeRequest:
      type: object
      required: [name, content]
      properties:
        name:
          type: string
          description: Item name (used as filename, `.md` appended automatically)
          example: auth-notes
        content:
          type: string
          description: Markdown content
          example: "# Auth Notes\n\nWe use JWT with 15min expiry..."
        projectDir:
          type: string
          description: Store in `<projectDir>/.antigravity/knowledge/` instead of global dir

    # Terminal
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            type:
              type: string
            code:
              type: string
  responses:
    BadRequest:
      description: Bad request — missing or invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              message: "`q` query parameter is required"
              type: invalid_request_error
...
    ServiceUnavailable:
      description: Antigravity IDE is not reachable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              message: "Antigravity not available: connection refused"
              type: service_unavailable
    UpstreamError:
      description: Upstream cascade error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              message: "StartCascade failed: timeout"
              type: upstream_error
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              message: "git worktree add failed: branch already exists"
              type: command_error